fix windows build; fix ansi-sgr for LS_COLORS

+++ src/common.h
@@ -108,7 +108,7 @@ ic_private ic_color_t ic_rgb(uint32_t hex);
 
 // Create a color from a 8-bit red/green/blue components.
 // The value of each component is capped between 0 and 255.
-ic_private ic_color_t ic_rgbx(int r, int g, int b);
+ic_private ic_color_t ic_rgbx(ssize_t r, ssize_t g, ssize_t b);
 
 #define IC_COLOR_NONE     (0)
 #define IC_RGB(rgb)       (0x1000000 | (uint32_t)(rgb)) // ic_rgb(rgb)  // define to it can be used as a constant
